+++ CHANGELOG.md
@@ -11,6 +11,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Adds telemetry hooks for connection start and stop events
 - Splits Sink.Connection.Client into two processes - one to bring up the connection, another to handle connection data. Also traps exits which should allow notifying SinkHandler when a connection errors. [CH46694]
 - Mox for testing
+- Server processes will send the `sink_handler` a message when a connection goes up/down.
+- Adds ServerConnectionHandler behavior
 
 
 ### Fixes